Split fromfile support into its own submodule (#20807)
Background: Option @fromfile paths are intended to be relative
to the buildroot, not the cwd. This will require us to plumb
buildroot information through in several places.

This change is prep work for that change, and is straightforward
and technical, despite the diff stats: It just splits fromfile support
out into its own submodule. This will make the subsequent changes
easier to grok.

The new fromfile.rs and fromfile_tests.rs contain code moved
from parser.rs and parser_tests.rs, with no functionality changes.
The other diffs are just import changes.
